+++ parser/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - The new `ReservedIdent` warning replaces the `ExpectedIdentFoundReserved` error.
+- Improve error message when an attribute is expected.
 
 ### Fixed